feat: add more dict functionality, more refactoring, fix bug with 'for (var item in list) {}' functionality, add lookahead functionality to the parser
This is quite a major commit that adds a bunch of different functionality.

Firstly, we expanded the parser to be able to lookahead. We initially only support a MAX_LOOKAHEAD value of 4 but this can be expanded as necessary.
We also add a 'peekToken()' method for peeking ahead.

Secondly, with this functionality, we fixed our loop iteration syntax. Rather than 'for (item in list) {}' we can
go back to 'for (var item in list) {}' which is more consistent with existing syntax.
This exposed some bugs in our opcode processing of OP_GREATER_EQUAL and OP_LESS_EQUAL where we weren't popping
values correctly. These were fixed but more debug logging added to the compiler.

Thirdly, we expanded our 'native' method functionality for dicts/lists by adding a base 'container' class.
ObjClasses' can now also have parents (though this is limited to the compiler for now) and this means that
lists and dicts can share common methods. Updated the native methods to support both lists and dicts.

Fourthly, added new native methods for dicts.