feat: add more language features (dicts + enums) (#11)
This is a large PR with a number of new features, bug fixes, and misc changes.

Firstly, we expanded the parser to be able to lookahead. We initially only support a MAX_LOOKAHEAD value of 4 but this can be expanded as necessary. We also add a 'peekToken()' method for peeking ahead.

Secondly, this corrects the list iteration syntax to:

for (var x in [1, 2, 3, 4]) {
    print(x);
}

This also fixes a bug in the OP_GREATER_EQUAL and OP_LESS_EQUAL vm handling.
Lots more debug printing was added for the debug build.

Thirdly, create a new base container class so that lists/dicts/other containers can share methods.

Fourthly, add dict support!

Finally, add enum support!

There was also more refactoring involved.